Refactor: extract PrefixSequencePolynomialHasher
Make hashing formulas abstract for the gapped k-mers algorithm
